Improve reading from a socket
First of all use flowing mode for reading by attaching a 'data' handler
to the readable stream (see https://nodejs.org/dist/latest-v8.x/docs/api/stream.html#stream_two_modes
for details).

Second allow to pass two functions to the read method. The first
function is a consumer function that creates data from read chunks. The
second function is intended to decide if reading has finished. If it
returns true or the function is missing, the promise is resolved with
the created data. If the consumer function is missing the read data is
concatenated by default.